LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/navigation-timing.https-expected.txt


FAIL Service worker controlled navigation timing assert_greater_than: expected a number greater than 0 but got 0
PASS Service worker controlled navigation timing network fallback
PASS Service worker controlled navigation timing redirect
